refactor: unify types, rename ambiguous files, split heavy barrels
- Replace GenImport/IESMImport with ScannedResolver as single source of truth
- Rename config.ts → create-config.ts, defaults.ts, file-header.ts, file-scanner.ts, runtime-generator.ts
- Split watcher/index.ts → create-watcher.ts + barrel
- Split pubsub/index.ts → memory-pubsub.ts + barrel
- Fix BASE_SCHEMA import (was importing from wrong module)
- Remove dead GenImport type alias and stale comments
