+++ src/core/codegen/plugins/index.ts
@@ -2,5 +2,5 @@
  * Codegen plugins barrel export
  */
 
-export { GENERATED_FILE_HEADER, pluginContent } from '../plugin'
+export { GENERATED_FILE_HEADER, pluginContent } from '../file-header'
 export { typedDocumentStringPlugin } from './typed-document-string'
