+++ src/core/create-config.ts
