+++ src/core/scanning/documents.ts
@@ -6,7 +6,7 @@
 import type { ScanContext, ScanResult } from '../types/scanning'
 import { relative } from 'pathe'
 import { GRAPHQL_GLOB_PATTERN } from '../constants'
-import { scanDirectory } from './common'
+import { scanDirectory } from './file-scanner'
 
 const REGEX_SPECIAL_CHARS_RE = /[.*+?^${}()|[\]\\]/g
