+++ src/core/scanning/file-scanner.ts
