+++ src/core/server/index.ts
@@ -9,6 +9,9 @@
  * - Standalone CLI (via srvx)
  */
 
+export type { CoreSecurityConfig } from '../types/config'
+
+export { APOLLO_USER_FACING_ERROR_CODES, createApolloServerInstance } from './apollo'
 // Sandbox script
 export { APOLLO_SANDBOX_CDN, createSandboxResponse, fetchSandboxScript } from './sandbox'
 
@@ -18,8 +21,5 @@ export type {
   CoreServerOptions,
   ServerFactory,
 } from './types'
-export type { CoreSecurityConfig } from '../types/config'
-
 // Server factories
 export { apolloSandboxHtml, createYogaServer } from './yoga'
-export { APOLLO_USER_FACING_ERROR_CODES, createApolloServerInstance } from './apollo'
