+++ src/nitro/routes/apollo-server.ts
@@ -1,10 +1,10 @@
 import type { BaseContext } from '@apollo/server'
+import type { EventHandler } from 'nitro/h3'
 import { importedConfig } from '#nitro-graphql/graphql-config'
 import { moduleConfig } from '#nitro-graphql/module-config'
 import { directives } from '#nitro-graphql/server-directives'
 import { resolvers } from '#nitro-graphql/server-resolvers'
 import { schemas } from '#nitro-graphql/server-schemas'
-import type { EventHandler } from 'nitro/h3'
 import { startServerAndCreateH3Handler } from 'nitro-graphql/apollo'
 import { defineEventHandler } from 'nitro/h3'
 import { BASE_SCHEMA_DEF } from '../../core/schema/builder'
