+++ src/nitro/types/define.ts
@@ -8,8 +8,13 @@ import type { ApolloServerOptions } from '@apollo/server'
 import type { YogaServerOptions } from 'graphql-yoga'
 import type { H3Event } from 'nitro/h3'
 
-// Re-export StandardSchemaV1 from core (canonical definition)
-export type { StandardSchemaV1 } from '../../core/types/standard-schema'
+// Re-export codegen types from core (authoritative definitions)
+export type {
+  ClientCodegenConfig,
+  ExternalServiceCodegenConfig,
+  SdkCodegenConfig,
+  ServerCodegenConfig,
+} from '../../core/types/codegen'
 
 export type Flatten<T> = T extends infer U ? { [K in keyof U]: U[K] } : never
 
@@ -19,6 +24,15 @@ export type DefineServerConfig<T extends NPMConfig = NPMConfig> = T['framework']
     ? Partial<ApolloServerOptions<H3Event>>
     : Partial<YogaServerOptions<H3Event, Partial<H3Event>>> | Partial<ApolloServerOptions<H3Event>>
 
+// Nitro-specific aliases for backward compatibility
+export type { ServerCodegenConfig as CodegenServerConfig } from '../../core/types/codegen'
+
+export type { ClientCodegenConfig as CodegenClientConfig } from '../../core/types/codegen'
+
+export type { SdkCodegenConfig as GenericSdkConfig } from '../../core/types/codegen'
+
+// Re-export SecurityConfig from core (single source of truth)
+export type { CoreSecurityConfig as SecurityConfig } from '../../core/types/config'
 // Re-export directive types from core (canonical definitions)
 export type {
   DefineDirectiveConfig,
@@ -30,19 +44,5 @@ export type {
   GraphQLBaseType,
   GraphQLScalarType,
 } from '../../core/types/define'
-
-// Re-export codegen types from core (authoritative definitions)
-export type {
-  ClientCodegenConfig,
-  ExternalServiceCodegenConfig,
-  SdkCodegenConfig,
-  ServerCodegenConfig,
-} from '../../core/types/codegen'
-
-// Re-export SecurityConfig from core (single source of truth)
-export type { CoreSecurityConfig as SecurityConfig } from '../../core/types/config'
-
-// Nitro-specific aliases for backward compatibility
-export type { ServerCodegenConfig as CodegenServerConfig } from '../../core/types/codegen'
-export type { ClientCodegenConfig as CodegenClientConfig } from '../../core/types/codegen'
-export type { SdkCodegenConfig as GenericSdkConfig } from '../../core/types/codegen'
+// Re-export StandardSchemaV1 from core (canonical definition)
+export type { StandardSchemaV1 } from '../../core/types/standard-schema'
