+++ tests/unit/scanning/schemas.test.ts
@@ -138,4 +138,3 @@ describe('scanSchemasCore', () => {
     })
   })
 })
-
